# Week8/server/openapi_server/models/product_update.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class ProductUpdate(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, name=None, price=None, owner_id=None):  # noqa: E501
        """ProductUpdate - a model defined in OpenAPI

        :param name: The name of this ProductUpdate.  # noqa: E501
        :type name: str
        :param price: The price of this ProductUpdate.  # noqa: E501
        :type price: float
        :param owner_id: The owner_id of this ProductUpdate.  # noqa: E501
        :type owner_id: int
        """
        self.openapi_types = {
            'name': str,
            'price': float,
            'owner_id': int
        }

        self.attribute_map = {
            'name': 'name',
            'price': 'price',
            'owner_id': 'ownerId'
        }

        self._name = name
        self._price = price
        self._owner_id = owner_id

    @classmethod
    def from_dict(cls, dikt) -> 'ProductUpdate':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ProductUpdate of this ProductUpdate.  # noqa: E501
        :rtype: ProductUpdate
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this ProductUpdate.


        :return: The name of this ProductUpdate.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this ProductUpdate.


        :param name: The name of this ProductUpdate.
        :type name: str
        """

        self._name = name

    @property
    def price(self) -> float:
        """Gets the price of this ProductUpdate.


        :return: The price of this ProductUpdate.
        :rtype: float
        """
        return self._price

    @price.setter
    def price(self, price: float):
        """Sets the price of this ProductUpdate.


        :param price: The price of this ProductUpdate.
        :type price: float
        """

        self._price = price

    @property
    def owner_id(self) -> int:
        """Gets the owner_id of this ProductUpdate.


        :return: The owner_id of this ProductUpdate.
        :rtype: int
        """
        return self._owner_id

    @owner_id.setter
    def owner_id(self, owner_id: int):
        """Sets the owner_id of this ProductUpdate.


        :param owner_id: The owner_id of this ProductUpdate.
        :type owner_id: int
        """

        self._owner_id = owner_id