# Week8/server/openapi_server/models/user_create.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class UserCreate(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, name=None, email=None):  # noqa: E501
        """UserCreate - a model defined in OpenAPI

        :param name: The name of this UserCreate.  # noqa: E501
        :type name: str
        :param email: The email of this UserCreate.  # noqa: E501
        :type email: str
        """
        self.openapi_types = {
            'name': str,
            'email': str
        }

        self.attribute_map = {
            'name': 'name',
            'email': 'email'
        }

        self._name = name
        self._email = email

    @classmethod
    def from_dict(cls, dikt) -> 'UserCreate':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UserCreate of this UserCreate.  # noqa: E501
        :rtype: UserCreate
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this UserCreate.


        :return: The name of this UserCreate.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this UserCreate.


        :param name: The name of this UserCreate.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def email(self) -> str:
        """Gets the email of this UserCreate.


        :return: The email of this UserCreate.
        :rtype: str
        """
        return self._email

    @email.setter
    def email(self, email: str):
        """Sets the email of this UserCreate.


        :param email: The email of this UserCreate.
        :type email: str
        """
        if email is None:
            raise ValueError("Invalid value for `email`, must not be `None`")  # noqa: E501

        self._email = email