# Week8/server/openapi_server/models/user_update.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class UserUpdate(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, name=None, email=None):  # noqa: E501
        """UserUpdate - a model defined in OpenAPI

        :param name: The name of this UserUpdate.  # noqa: E501
        :type name: str
        :param email: The email of this UserUpdate.  # noqa: E501
        :type email: str
        """
        self.openapi_types = {
            'name': str,
            'email': str
        }

        self.attribute_map = {
            'name': 'name',
            'email': 'email'
        }

        self._name = name
        self._email = email

    @classmethod
    def from_dict(cls, dikt) -> 'UserUpdate':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UserUpdate of this UserUpdate.  # noqa: E501
        :rtype: UserUpdate
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this UserUpdate.


        :return: The name of this UserUpdate.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this UserUpdate.


        :param name: The name of this UserUpdate.
        :type name: str
        """

        self._name = name

    @property
    def email(self) -> str:
        """Gets the email of this UserUpdate.


        :return: The email of this UserUpdate.
        :rtype: str
        """
        return self._email

    @email.setter
    def email(self, email: str):
        """Sets the email of this UserUpdate.


        :param email: The email of this UserUpdate.
        :type email: str
        """

        self._email = email